KModes demo
A student asked about "clustering categorical variables". I don't entirely agree with using KModes over my own intuition (with most datasets where features are coherent, but I ran a test to demonstrate what it can do. It can be useful if there is a large featureset, but I still don't like how inexact the results are- it's a big risk to take.